DRFBOT/apps/products/serializers.py
from rest_framework import serializers

from apps.products.models import Product


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ("id", "title", "descriptions", "price", "created_at")
